Clean up of Home screen cards and extracted JetlaggedTimeGraph into separate file.
