+++ JetLagged/app/src/main/java/com/example/jetlagged/JetLaggedScreen.kt
@@ -19,58 +19,32 @@
 package com.example.jetlagged
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.FlowColumn
 import androidx.compose.foundation.layout.FlowRow
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.widthIn
-import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.material3.windowsizeclass.WindowWidthSizeClass
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.movableContentOf
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.drawBehind
-import androidx.compose.ui.geometry.CornerRadius
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.LookaheadScope
-import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import com.example.jetlagged.ui.theme.HeadingStyle
-import com.example.jetlagged.ui.theme.SmallHeadingStyle
-import com.example.jetlagged.ui.theme.Yellow
-import com.example.jetlagged.ui.theme.YellowVariant
-import java.time.DayOfWeek
-import java.time.format.TextStyle
-import java.util.Locale
 
 @OptIn(ExperimentalLayoutApi::class, ExperimentalComposeUiApi::class)
 @MultiDevicePreview
 @Composable
-fun JetLaggedScreen(windowSizeClass: WindowWidthSizeClass = WindowWidthSizeClass.Compact,
-                    viewModel: JetLaggedHomeScreenViewModel = JetLaggedHomeScreenViewModel() // todo fix this to come from nav
-                    ) {
+fun JetLaggedScreen(
+    windowSizeClass: WindowWidthSizeClass = WindowWidthSizeClass.Compact,
+    viewModel: JetLaggedHomeScreenViewModel = JetLaggedHomeScreenViewModel() // todo fix this to come from nav
+) {
     Column(
         modifier = Modifier
             .fillMaxSize()
@@ -81,7 +55,8 @@ fun JetLaggedScreen(windowSizeClass: WindowWidthSizeClass = WindowWidthSizeClass
             JetLaggedHeader(modifier = Modifier.fillMaxWidth())
         }
 
-        val uiState = viewModel.uiState.collectAsState() // todo do i change to collectAsStateWithLifecycle
+        val uiState =
+            viewModel.uiState.collectAsState() // todo do i change to collectAsStateWithLifecycle
         val timeSleepSummaryCards = remember {
             movableContentOf {
                 AverageTimeInBedCard(modifier = Modifier)
@@ -113,110 +88,8 @@ fun JetLaggedScreen(windowSizeClass: WindowWidthSizeClass = WindowWidthSizeClass
                     }
                 }
                 AmbienceCard()
+                RoomTemperature()
             }
-        }
-    }
-}
-
-@Composable
-private fun JetLaggedSleepGraphCard(
-    sleepState: SleepGraphData
-) {
-    var selectedTab by remember { mutableStateOf(SleepTab.Week) }
-
-    BasicInformationalCard(
-        borderColor = Yellow,
-        modifier = Modifier.animateBounds(Modifier.widthIn(max = 400.dp))
-    ) {
-        Column {
-            HomeScreenCardHeading(text = "Sleep")
-            JetLaggedHeaderTabs(
-                onTabSelected = { selectedTab = it },
-                selectedTab = selectedTab,
-                modifier = Modifier.padding(top = 16.dp)
-            )
-            Spacer(modifier = Modifier.height(16.dp))
-            JetLaggedTimeGraph(
-                sleepState
-            )
-        }
-    }
-}
-
-@Composable
-private fun JetLaggedTimeGraph(
-    sleepGraphData: SleepGraphData,
-    modifier: Modifier = Modifier
-) {
-    val scrollState = rememberScrollState()
-
-    val hours = (sleepGraphData.earliestStartHour..23) + (0..sleepGraphData.latestEndHour)
-
-    TimeGraph(
-        modifier = modifier
-            .horizontalScroll(scrollState)
-            .wrapContentSize(),
-        dayItemsCount = sleepGraphData.sleepDayData.size,
-        hoursHeader = {
-            HoursHeader(hours)
-        },
-        dayLabel = { index ->
-            val data = sleepGraphData.sleepDayData[index]
-            DayLabel(data.startDate.dayOfWeek)
-        },
-        bar = { index ->
-            val data = sleepGraphData.sleepDayData[index]
-            // We have access to Modifier.timeGraphBar() as we are now in TimeGraphScope
-            SleepBar(
-                sleepData = data,
-                modifier = Modifier
-                    .padding(bottom = 8.dp)
-                    .timeGraphBar(
-                        start = data.firstSleepStart,
-                        end = data.lastSleepEnd,
-                        hours = hours,
-                    )
-            )
-        }
-    )
-}
-
-@Composable
-private fun DayLabel(dayOfWeek: DayOfWeek) {
-    Text(
-        dayOfWeek.getDisplayName(
-            TextStyle.SHORT, Locale.getDefault()
-        ),
-        Modifier
-            .height(24.dp)
-            .padding(start = 8.dp, end = 24.dp),
-        style = SmallHeadingStyle,
-        textAlign = TextAlign.Center
-    )
-}
-
-@Composable
-private fun HoursHeader(hours: List<Int>) {
-    Row(
-        Modifier
-            .padding(bottom = 16.dp)
-            .drawBehind {
-                val brush = Brush.linearGradient(listOf(YellowVariant, Yellow))
-                drawRoundRect(
-                    brush,
-                    cornerRadius = CornerRadius(10.dp.toPx(), 10.dp.toPx()),
-                )
-            }
-    ) {
-        hours.forEach {
-            Text(
-                text = "$it",
-                textAlign = TextAlign.Center,
-                modifier = Modifier
-                    .width(50.dp)
-                    .padding(vertical = 4.dp),
-                style = SmallHeadingStyle
-            )
         }
     }
 }
